refactor: 레거시 코드를 삭제해도 서버가 돌아가도록 리팩토링
