+++ mvc/src/test/java/nextstep/mvc/DispatcherServletTest.java
@@ -4,7 +4,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 
-import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
+import nextstep.mvc.controller.AnnotationHandlerMapping;
 import org.junit.jupiter.api.Test;
 
 class DispatcherServletTest {
